+++ worker/__init__.py
@@ -1,5 +1,4 @@
 # ratchet.worker
 from worker.log_buffer import LogBuffer, LogEntry
-from worker.service import WorkerService, WorkerSettings
 
-__all__ = ["LogBuffer", "LogEntry", "WorkerService", "WorkerSettings"]
+__all__ = ["LogBuffer", "LogEntry"]